added 104-fibonnaci.c

+++ functions_nested_loops/104-fibonacci.c
@@ -7,7 +7,7 @@
 
 int	main(void)
 {
-	float	number1 = 1, number2 = 2, number3;
+	double	number1 = 1, number2 = 2, number3;
 	int	index = 3;
 
 	printf("%.0f, %.0f", number1, number2);
